Update: Integration of json_interface to UI
The Path to file now needs to be specified beforehand, and may the file may be created or an existing one can be chosen.
Changed the used interface and logic of a few functions to properly communicate with user.

TODO: Show user that path to file is invalid upon failure